+++ changelog.md
@@ -5,6 +5,7 @@
   * Drops from zombies
   * Required to make a bed
   * Can be used to make Chainmail armor
+* Cleansed lapis now has greatly increased chance to drop
 
 ## 3.11.3
 * Heavily reduced armor slowdown